+++ backend/src/profile/profile.module.ts
@@ -5,10 +5,11 @@ import { SequelizeModule } from "@nestjs/sequelize";
 import { UserModel } from "../user/models/user.model";
 import { FollowModel } from "../user/models/follow.model";
 import { FileModule } from "../file/file.module";
+import { PhotoModel } from "../user/models/photo.model";
 
 @Module({
   controllers: [ProfileController],
   providers: [ProfileService],
-  imports: [SequelizeModule.forFeature([UserModel, FollowModel]), FileModule],
+  imports: [SequelizeModule.forFeature([UserModel, FollowModel, PhotoModel]), FileModule],
 })
 export class ProfileModule {}